capsule-manager-sdk/python/requirements.txt
cryptography>=41.0.2
grpcio==1.62.1
ruamel.yaml==0.18.6
click==8.1.7
trustflow-verification==0.3.0.dev20240801